+++ packages/components/src/templates/next/components/internal/SearchableTable/DGS/DGSSearchableTable.tsx
