+++ packages/components/src/templates/next/components/internal/SearchableTable/DGS/index.ts
@@ -1,1 +1,1 @@
-export { DGSSearchableTable } from "./SearchableTable"
+export { DGSSearchableTable } from "./DGSSearchableTable"
