+++ packages/components/src/templates/next/components/internal/SearchableTable/Native/index.ts
@@ -1,1 +1,1 @@
-export { default } from "./SearchableTable"
+export { NativeSearchableTable } from "./NativeSearchableTable"
